+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Correct logging for Ambient Light mode active skip
+
 ## 3.0.0-alpha.10 (2024-03-17)
 
 * (theimo1221) Reduce duplicate logging